fix: novadev-693
