+++ src/cli.ts
@@ -1,10 +1,10 @@
 import Yargs from 'yargs/yargs';
 import YargsCommandBuilderOptions from './common/yargs/yargs-command-builder-options';
-import { configureCommandOptions, readConfigFile } from './commands/configure';
+import { readConfigFile } from './commands/configure';
 import { Arguments, Argv } from 'yargs';
 import errorHandler from './error-handler';
 
-const readConfig = (configFile: string): object => {
+export const readConfig = (configFile: string): object => {
   return readConfigFile(configFile, process.argv[2] === 'configure');
 };
 
@@ -26,8 +26,6 @@ const configureYargs = (yargInstance: Argv): Promise<Arguments> => {
       };
       const argv = await yargInstance
         .scriptName('dc-cli')
-        .options(configureCommandOptions)
-        .config('config', readConfig)
         .commandDir('./commands', YargsCommandBuilderOptions)
         .strict()
         .demandCommand(1, 'Please specify at least one command')
