+++ src/commands/content-repository.ts
@@ -1,5 +1,7 @@
 import { Argv } from 'yargs';
+import { readConfig } from '../cli';
 import YargsCommandBuilderOptions from '../common/yargs/yargs-command-builder-options';
+import { configureCommandOptions } from './configure';
 
 export const command = 'content-repository';
 
@@ -8,6 +10,8 @@ export const desc = 'Content Repository';
 export const builder = (yargs: Argv): Argv =>
   yargs
     .commandDir('content-repository', YargsCommandBuilderOptions)
+    .options(configureCommandOptions)
+    .config('config', readConfig)
     .demandCommand()
     .help();
