+++ src/commands/extension.ts
@@ -1,5 +1,7 @@
 import { Argv } from 'yargs';
+import { readConfig } from '../cli';
 import YargsCommandBuilderOptions from '../common/yargs/yargs-command-builder-options';
+import { configureCommandOptions } from './configure';
 
 export const command = 'extension';
 
@@ -8,5 +10,7 @@ export const desc = 'Extension';
 export const builder = (yargs: Argv): Argv =>
   yargs
     .commandDir('extension', YargsCommandBuilderOptions)
+    .options(configureCommandOptions)
+    .config('config', readConfig)
     .demandCommand()
     .help();
