+++ src/commands/hub.ts
@@ -1,9 +1,16 @@
 import { Argv } from 'yargs';
 import YargsCommandBuilderOptions from '../common/yargs/yargs-command-builder-options';
 import ConfigManager from '../common/hub-manager';
+import { CommandOptions } from '../interfaces/command-options.interface';
 
 export const command = 'hub';
 export const desc = 'Hub';
+
+const commandOptions: CommandOptions = {
+  clientId: { type: 'string' },
+  clientSecret: { type: 'string' },
+  hubId: { type: 'string' }
+};
 
 export const hubBuilder = (yargs: Argv): Argv =>
   yargs.positional('hub', {
@@ -19,7 +26,7 @@ export const builder = (yargs: Argv): Argv =>
   yargs
     .commandDir('hub', YargsCommandBuilderOptions)
     .demandCommand()
-    .command('add', 'Add hub', ConfigManager.addHub)
+    .command('add', 'Add hub', commandOptions, ConfigManager.addHub)
     .command('list', 'List hubs', ConfigManager.listHubs)
     .command('ls', 'List hubs', ConfigManager.listHubs)
     .command('use [hub]', 'Use hub', hubBuilder, ConfigManager.useHub)
